refactor: nightly maintainability and docs
Move the pure VideoLayout projections out of the UI layer. The scene-segment
module lived under app/components/video as a "use hook" file while holding only
domain logic: buildSceneSegments, buildSequenceIndex, findSceneSequence and
findSegmentIndexAtFrame take a VideoLayout and return plain data, and
SceneSegment is a domain shape that pulled its SeekThumbnail type out of a
tooltip component -- lib depending on app, inverted.

lib/video/sceneSegments.ts is now their one home, and SeekTooltip imports the
thumbnail type from it. The one React binding, useActiveSegmentIndex, moves to
its own leaf next to usePlayerState so each module has a single purpose.
