+++ app/components/video/BottomTransportBar.tsx
@@ -5,11 +5,11 @@ import { ChevronsLeft, ChevronsRight, Pause, Play } from "@/components/ui/icon";
 import { TooltipIconButton } from "@/components/ui/icon-button";
 import { cn } from "@/lib/utils";
 import { toFrames } from "@/lib/video/frames";
+import { findSegmentIndexAtFrame } from "@/lib/video/sceneSegments";
 import { BottomViewToggle } from "./BottomViewToggle";
 import { usePlayerControl } from "./PlayerControlContext";
 import { usePlayerPosition } from "./PlayerPositionContext";
 import { useLayout } from "./VideoLayoutContext";
-import { findSegmentIndexAtFrame } from "./useSceneSegments";
 import { usePlayerPlaying } from "./usePlayerState";
 import { SCRUB_BAR_HEIGHT } from "./ScrubBar";
 import { SegmentedSeekBar } from "./SegmentedSeekBar";
